+++ src/config/get.c
@@ -679,7 +679,18 @@ void fs_x509_generate_conf(void *funcarg, void* arg)
 
 		for (uint64_t i = 0; i < tls_fs->match->l; ++i) {
 			json_t *token = json_string(tls_fs->match->str[i]->s);
-			json_array_object_insert(ctx, "", token);
+			json_array_object_insert(match, "", token);
+		}
+	}
+
+	if (tls_fs->except)
+	{
+		json_t *except = json_array();
+		json_array_object_insert(ctx, "except", except);
+
+		for (uint64_t i = 0; i < tls_fs->except->l; ++i) {
+			json_t *token = json_string(tls_fs->except->str[i]->s);
+			json_array_object_insert(except, "", token);
 		}
 	}